requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.6
decorator==5.1.1
Django==4.2
django-cors-headers==4.0.0
django-jazzmin==2.6.0
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
dynaconf==3.1.12
exceptiongroup==1.1.1
executing==1.2.0
hvac==1.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
ipython==8.14.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
loguru==0.7.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
peewee==3.16.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
prettytable==3.7.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyhcl==0.4.4
PyJWT==2.6.0
pytest==7.3.1
pytest-cov==2.12.1
python-decouple==3.8
pytz==2023.3
rand==0.3.0
readonly==0.1.2
requests==2.30.0
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.17.29
ruamel.yaml.clib==0.2.7
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tokenize-rt==5.0.0
toml==0.10.2
tomli==2.0.1
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
